requirements.txt
click>=6.7
moto>=1.3.7
numpy>=1.16.4
pandas>=0.24.0
nltk>=3.2.5
pydot>=1.4.1
networkx>=2.1
dill>=0.2.8.2
pytest~=3.10.1
testfixtures>=6.8.2
boto3>=1.9.0
matplotlib>=2.1.2
psycopg2_binary>=2.8.2
protobuf>=3.8.0
mysql-connector-python>=8.0.16
psycopg2>=2.8.3
pytest-cov>=2.6.0
scikit-learn>=0.21.2
atlas>=0.27.0
google-cloud-storage>=1.16.0
pygraphviz>=1.5
jstyleson==0.0.2